Core/Misc: First set of required db changes for 9.0
Port From (https://github.com/TrinityCore/TrinityCore/commit/4ca29f3cb050896a8a6c1fea5b2364b2b11c22cd)

+++ Source/Game/Entities/Creature/Gossip.cs
@@ -274,11 +274,11 @@ namespace Game.Misc
                 if (quest != null)
                 {
                     ClientGossipText text = new ClientGossipText();
-                    text.QuestID = (int)questID;
+                    text.QuestID = questID;
                     text.ContentTuningID = quest.ContentTuningId;
                     text.QuestType = item.QuestIcon;
-                    text.QuestFlags = (int)quest.Flags;
-                    text.QuestFlagsEx = (int)quest.FlagsEx;
+                    text.QuestFlags = (uint)quest.Flags;
+                    text.QuestFlagsEx = (uint)quest.FlagsEx;
                     text.Repeatable = quest.IsRepeatable();
 
                     text.QuestTitle = quest.LogTitle;
